+++ src/main/java/com/comprehensive/eureka/plan/entity/Plan.java
@@ -22,7 +22,7 @@ public class Plan {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "plan_id")
-    private Integer planId;
+    private Long planId;
 
     private String planName;
