+++ src/main/java/com/comprehensive/eureka/plan/entity/VoiceCall.java
@@ -20,7 +20,7 @@ public class VoiceCall {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "voice_call_id")
-    private Integer voiceCallId;
+    private Long voiceCallId;
 
     private Integer voiceAllowance;
